implement POST methods for Create and Edit Views

+++ ViewModels/Terraria/Boss/BossCreateViewModel.cs
@@ -31,7 +31,7 @@ namespace TerrariaDB.ViewModels.Terraria.Boss
         public string Sprite { get; set; } = string.Empty;
         public int Hp { get; set; }
         public int Defense { get; set; }
-        public int EntityId { get; set; }
+        public short EntityId { get; set; }
         public int ContactDamage { get; set; }
         public List<BossStageEnemyCreateViewModel> SpawnedEnemies { get; set; } = new();
         public List<BossStageDropCreateViewModel> Drops { get; set; } = new();
