+++ ViewModels/Terraria/Recipe/RecipeCreateViewModel.cs
@@ -5,7 +5,7 @@ namespace TerrariaDB.ViewModels.Terraria.Recipe
     public class RecipeCreateViewModel
     {
         public string ResultItemId { get; set; } = string.Empty;
-        public int ResultItemQuantity { get; set; }
+        public short ResultItemQuantity { get; set; }
         public string? CraftingStationName { get; set; }
         public List<RecipeCreateIngredientViewModel> Ingredients { get; set; } = new();
         public List<SelectListItem> AvailableItems { get; set; } = new();
